Moved helper/xcodeproj_ext.rb to core_ext/xcodeproj.rb. Moved some things from IOSHelper to ToolHelper.
