+++ lib/branch_io_cli/core_ext.rb
@@ -1,2 +1,3 @@
 require "branch_io_cli/core_ext/io.rb"
 require "branch_io_cli/core_ext/regexp.rb"
+require "branch_io_cli/core_ext/xcodeproj.rb"
