+++ lib/branch_io_cli/core_ext/xcodeproj.rb
